requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Google Cloud / Vertex AI / Gemini
google-cloud-storage==2.18.2
google-cloud-firestore==2.20.0
google-cloud-aiplatform==1.63.0
google-genai==0.5.0

# Core numerics and CV
numpy<2.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
mediapipe==0.10.14
scikit-image==0.24.0

# PyTorch ecosystem (ViT + Grad-CAM)
torch==2.2.2
timm==1.0.9
grad-cam>=1.4.8

# TensorFlow ecosystem (Bayesian fusion)
tensorflow==2.15.0
tensorflow-probability==0.23.0

# CLIP (OpenAI official)
git+https://github.com/openai/CLIP.git

# Server
gunicorn==22.0.0
